Update to 8.0.1 (2428)
 Conflicts:
	TMessagesProj/src/main/AndroidManifest.xml
	TMessagesProj/src/main/java/org/telegram/messenger/AndroidUtilities.java
	TMessagesProj/src/main/java/org/telegram/messenger/AutoMessageReplyReceiver.java
	TMessagesProj/src/main/java/org/telegram/messenger/BuildVars.java
	TMessagesProj/src/main/java/org/telegram/messenger/GoogleVoiceClientService.java
	TMessagesProj/src/main/java/org/telegram/messenger/MessagesController.java
	TMessagesProj/src/main/java/org/telegram/messenger/MessagesStorage.java
	TMessagesProj/src/main/java/org/telegram/messenger/MusicBrowserService.java
	TMessagesProj/src/main/java/org/telegram/messenger/NotificationsController.java
	TMessagesProj/src/main/java/org/telegram/messenger/SegmentTree.java
	TMessagesProj/src/main/java/org/telegram/messenger/SharedConfig.java
	TMessagesProj/src/main/java/org/telegram/messenger/SmsReceiver.java
	TMessagesProj/src/main/java/org/telegram/messenger/WearDataLayerListenerService.java
	TMessagesProj/src/main/java/org/telegram/tgnet/NativeByteBuffer.java
	TMessagesProj/src/main/java/org/telegram/ui/ActionBar/ActionBarLayout.java
	TMessagesProj/src/main/java/org/telegram/ui/ActionBar/BaseFragment.java
	TMessagesProj/src/main/java/org/telegram/ui/Adapters/ContactsAdapter.java
	TMessagesProj/src/main/java/org/telegram/ui/Adapters/DialogsSearchAdapter.java
	TMessagesProj/src/main/java/org/telegram/ui/Adapters/SearchAdapter.java
	TMessagesProj/src/main/java/org/telegram/ui/ArchivedStickersActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/Cells/DialogCell.java
	TMessagesProj/src/main/java/org/telegram/ui/ChannelAdminLogActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/ChatActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/AdminLogFilterAlert.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/AlertsCreator.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/AudioPlayerAlert.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/BlockingUpdateView.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/ChatActivityEnterView.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/ChatAttachAlertContactsLayout.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/EditTextEmoji.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/FragmentContextView.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/GroupVoipInviteAlert.java
	TMessagesProj/src/main/java/org/telegram/ui/Components/ProfileGalleryView.java
	TMessagesProj/src/main/java/org/telegram/ui/FilterUsersActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/GroupCallActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/GroupCreateActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/LaunchActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/LocationActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/NotificationsSettingsActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/PaymentFormActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/PhotoViewer.java
	TMessagesProj/src/main/java/org/telegram/ui/PrivacyControlActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/ProfileActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/ProfileNotificationsActivity.java
	TMessagesProj/src/main/java/org/telegram/ui/ThemePreviewActivity.java

+++ TMessagesProj/src/main/java/org/telegram/messenger/voip/VoIPService.java
@@ -488,7 +488,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 	public boolean mutedByAdmin() {
 		ChatObject.Call call = groupCall;
 		if (call != null) {
-			int selfId = getSelfId();
+			long selfId = getSelfId();
 			TLRPC.TL_groupCallParticipant participant = call.participants.get(selfId);
             return participant != null && !participant.can_self_unmute && participant.muted && !ChatObject.canManageCalls(chat);
 		}
@@ -669,14 +669,14 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 			throw new IllegalStateException("No account specified when starting VoIP service");
 		}
 		classGuid = ConnectionsManager.generateClassGuid();
-		int userID = intent.getIntExtra("user_id", 0);
-		int chatID = intent.getIntExtra("chat_id", 0);
+		long userID = intent.getLongExtra("user_id", 0);
+		long chatID = intent.getLongExtra("chat_id", 0);
 		createGroupCall = intent.getBooleanExtra("createGroupCall", false);
 		hasFewPeers = intent.getBooleanExtra("hasFewPeers", false);
 		joinHash = intent.getStringExtra("hash");
-		int peerChannelId = intent.getIntExtra("peerChannelId", 0);
-		int peerChatId = intent.getIntExtra("peerChatId", 0);
-		int peerUserId = intent.getIntExtra("peerUserId", 0);
+		long peerChannelId = intent.getLongExtra("peerChannelId", 0);
+		long peerChatId = intent.getLongExtra("peerChatId", 0);
+		long peerUserId = intent.getLongExtra("peerUserId", 0);
 		if (peerChatId != 0) {
 			groupCallPeer = new TLRPC.TL_inputPeerChat();
 			groupCallPeer.chat_id = peerChatId;
@@ -839,7 +839,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		createGroupInstance(CAPTURE_DEVICE_CAMERA, false);
 	}
 
-	public int getCallerId() {
+	public long getCallerId() {
 		if (user != null) {
 			return user.id;
 		} else {
@@ -1324,7 +1324,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		tgVoip[CAPTURE_DEVICE_CAMERA].onSignalingDataReceive(data.data);
 	}
 
-	public int getSelfId() {
+	public long getSelfId() {
 		if (groupCallPeer == null) {
 			return UserConfig.getInstance(currentAccount).clientUserId;
 		}
@@ -1341,7 +1341,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		if (chat == null || groupCall == null || groupCall.call.id != update.call.id) {
 			return;
 		}
-		int selfId = getSelfId();
+		long selfId = getSelfId();
 		for (int a = 0, N = update.participants.size(); a < N; a++) {
 			TLRPC.TL_groupCallParticipant participant = update.participants.get(a);
 			if (participant.left) {
@@ -1751,7 +1751,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 				if (response != null) {
 					AndroidUtilities.runOnUIThread(() -> mySource[CAPTURE_DEVICE_CAMERA] = ssrc);
 					TLRPC.Updates updates = (TLRPC.Updates) response;
-					int selfId = getSelfId();
+					long selfId = getSelfId();
 					for (int a = 0, N = updates.updates.size(); a < N; a++) {
 						TLRPC.Update update = updates.updates.get(a);
 						if (update instanceof TLRPC.TL_updateGroupCallParticipants) {
@@ -1820,7 +1820,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 				TLRPC.Updates updates = (TLRPC.Updates) response;
 				AndroidUtilities.runOnUIThread(() -> {
 					if (tgVoip[CAPTURE_DEVICE_SCREEN] != null) {
-						int selfId = getSelfId();
+						long selfId = getSelfId();
 						for (int a = 0, N = updates.updates.size(); a < N; a++) {
 							TLRPC.Update update = updates.updates.get(a);
 							if (update instanceof TLRPC.TL_updateGroupCallConnection) {
@@ -1967,7 +1967,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		if (groupCall == null || tgVoip[CAPTURE_DEVICE_CAMERA] == null) {
 			return;
 		}
-		int selfId = getSelfId();
+		long selfId = getSelfId();
 		ArrayList<RequestedParticipant> participants = null;
 		for (int a = 0, N = unknown.length; a < N; a++) {
 			TLRPC.TL_groupCallParticipant p = groupCall.participantsBySources.get(unknown[a]);
@@ -2852,7 +2852,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		startForeground(ID_ONGOING_CALL_NOTIFICATION, builder.getNotification());
 	}
 
-	private void startRingtoneAndVibration(int chatID) {
+	private void startRingtoneAndVibration(long chatID) {
 		SharedPreferences prefs = MessagesController.getNotificationsSettings(currentAccount);
 		AudioManager am = (AudioManager) getSystemService(AUDIO_SERVICE);
 		boolean needRing = am.getRingerMode() != AudioManager.RINGER_MODE_SILENT;
